Add block test builder
