+++ shared_model/builders/protobuf/builder_templates/query_template.hpp
@@ -28,6 +28,13 @@
 namespace shared_model {
   namespace proto {
 
+  /**
+     * Template query builder for creating new types of query builders by
+     * means of replacing template parameters
+     * @tparam S -- field counter for checking that all required fields are set
+     * @tparam SV -- stateless validator called when build method is invoked
+     * @tparam BT -- build type of built object returned by build method
+     */
     template <int S = 0, typename SV = validation::DefaultQueryValidator, typename BT = UnsignedWrapper<Query>>
     class TemplateQueryBuilder {
      private:
